+++ PiwikPROSDK.xcframework/ios-arm64_x86_64-simulator/PiwikPROSDK.framework/Headers/PiwikPROSDK.h
@@ -16,6 +16,8 @@ FOUNDATION_EXPORT const unsigned char PiwikPROSDKVersionString[];
 #import "PiwikTransactionItem.h"
 #import "PiwikTransactionBuilder.h"
 #import "PiwikTrackedViewController.h"
+#import "Product.h"
+#import "EcommerceProducts.h"
 
 // Third party licenses
 //